requirements.txt
altgraph==0.17.4
ffmpeg-python==0.2.0
future==1.0.0
MouseInfo==0.1.3
packaging==24.0
pefile==2023.2.7
pillow==10.3.0
PyAutoGUI==0.9.54
pydub==0.25.1
pygame==2.5.2
PyGetWindow==0.0.9
pyinstaller==6.6.0
pyinstaller-hooks-contrib==2024.6
PyMsgBox==1.0.9
PyPDF2==3.0.1
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
pytweening==1.2.0
pywin32-ctypes==0.2.2
setuptools==69.5.1
